+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Unreleased changes
 ------------------
 
+
+Version 7
+---------
+
+7.7.0 / 2017-10-11
+~~~~~~~~~~~~~~~~~~
+* (aanet/tohd5) new option to read run ID from header, not event. 
+  in old versions of JTE, the event.run_id is overwritten with the default, 1.
+* there is now a new command line utility called ``streamds`` for non-pythonistas
+* The new ``km3pipe.ahrs`` now contains AHRS calibration routines
+
+
 7.6.1 / 2017-10-09
 ~~~~~~~~~~~~~~~~~~
 * ``HDF5Sink`` now uses the new ``HDF5MetaData`` class two write more verbose
@@ -20,9 +32,6 @@ Unreleased changes
 ~~~~~~~~~~~~~~~~~~
 * Option to Ignore hits in pumps
 * fix aanet fitinf enum
-
-Version 7
----------
 
 7.5.4 / 2017-09-25
 ~~~~~~~~~~~~~~~~~~